Stop including shared library in java_wrap_cc runtime_deps.
This is a breaking change. It includes a Java util for loading a library from JAR resources.

The native library is included as a dep of the resulting java_library. According to the Bazel docs, this should result in a native shared library being included in the resulting JAR. According to bazelbuild/bazel#3079, this does not currently function as documented. Therefore, the macro generates a shared library target that the caller can include in JAR resources.